backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.31.0
aiosqlite==0.22.1
redis==5.0.1
hiredis==3.3.1
openai==1.3.7
gTTS==2.5.4
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.3.0
websockets==12.0
aiofiles==23.2.1
faster-whisper==1.0.3
transformers==4.35.2
sentencepiece==0.1.99
langdetect==1.0.9
torch==2.1.1 --index-url https://download.pytorch.org/whl/cpu
torchaudio==2.1.1 --index-url https://download.pytorch.org/whl/cpu
